add notes about where the bugs are but make no attempt to fix them
This is because I could not see where each of the "a" elements hold
info on the user name, and I don't care because I get the point now:
you can iterate over the elements and find things.

The "hide sidebar" section does work: It does hide the sidebar on the
right side of the page.